fix sort pokemon use case
