+++ src/features/pokemon-list/application/use-cases/sort-pokemon-list-by-height/SortPokemonLIstByHeightUseCase.ts
@@ -1,0 +1,7 @@
+import { PokemonListItem } from "../../../domain/entities/PokemonListItem";
+
+export class SortPokemonsByHeightUseCase {
+  static execute(list: PokemonListItem[]): PokemonListItem[] {
+    return [...list].sort((a, b) => a.height - b.height);
+  }
+}
