--- src/features/pokemon-list/use-cases/get-pokemon-list/sort-pokemon-list-by-height/SortPokemonLIstByHeightUseCase.ts
@@ -1,7 +1,0 @@
-import { Pokemon } from "../../../domain/entities/Pokemon";
-
-export class SortPokemonsByHeightUseCase {
-  static execute(list: Pokemon[]): Pokemon[] {
-    return [...list].sort((a, b) => a.height - b.height);
-  }
-}
